# docs/contracts/openapi/generated.json
{
  "components": {
    "schemas": {
...
      "FieldError": {
        "additionalProperties": false,
        "description": "One field-level failure inside a problem response.\n\nPresent so that a rejected submission can be rendered against the form that produced\nit — FR-PLAT-11 requires deterministic failures to name the offending field rather\nthan only the operation.",
        "properties": {
          "code": {
            "description": "Stable machine code for this field.",
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "field": {
            "description": "Dotted path to the offending field, e.g. 'spec.family'.",
            "title": "Field",
            "type": "string"
          },
          "message": {
            "description": "Human-readable explanation.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "JobKind": {
        "description": "The typed operation a Job performs (`07` §2, job.schema.json).",
        "enum": [
          "dataset.ingest",
          "dataset.validate",
          "dataset.profile",
          "dataset.derive",
          "model.fit",
          "model.transparency",
          "model.backtest",
          "model.compare",
          "objective.certify",
          "rating.compile",
          "rating.regression",
          "score.batch",
          "dislocation.run",
          "optimisation.run",
          "gipp.check",
          "monitor.run",
          "dossier.generate",
          "export.regulatory",
          "blob.gc"
        ],
        "title": "JobKind",
        "type": "string"
      },
...
      "JobStatus": {
        "enum": [
          "queued",
          "running",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "Page_JobLogLine_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobLogLine"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque; pass back verbatim. Null when this is the last page.",
            "title": "Next Cursor"
          },
          "total_estimate": {
            "description": "Matching rows, counted up to 10000. An estimate by design.",
            "minimum": 0.0,
            "title": "Total Estimate",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total_estimate"
        ],
        "title": "Page[JobLogLine]",
        "type": "object"
      },
      "Page_Job_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque; pass back verbatim. Null when this is the last page.",
            "title": "Next Cursor"
          },
          "total_estimate": {
            "description": "Matching rows, counted up to 10000. An estimate by design.",
            "minimum": 0.0,
            "title": "Total Estimate",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total_estimate"
        ],
        "title": "Page[Job]",
        "type": "object"
      },
...
      "ProblemDetail": {
        "additionalProperties": false,
        "description": "An RFC 9457 problem response.\n\n`code` is the contract, not `title` or `status`: titles are prose and several\nconditions share a status, so a client that branches on anything else is branching on\nsomething the platform is free to reword.",
        "properties": {
          "code": {
            "description": "Stable machine-readable code, namespaced per module and enumerated in that module's spec. This is what clients branch on.",
            "pattern": "^[A-Z][A-Z0-9_]*$",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explanation specific to this occurrence.",
            "title": "Detail"
          },
          "errors": {
            "default": [],
            "description": "Field-level errors, where the failure is field-attributable.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path of the failing request.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code.",
            "maximum": 599.0,
            "minimum": 100.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short human-readable summary, stable per type.",
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "OpenTelemetry trace id for this request. Appears in every log line for the request, so a support conversation starts with an identifier (FR-PLAT-42).",
            "title": "Trace Id"
          },
          "type": {
            "default": "about:blank",
            "description": "URI identifying the problem type; documentation lives at that URI.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "GI Pricing Platform API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/jobs": {
      "get": {
        "description": "Filtered, cursor-paginated (`00` §5.2).",
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "submitted_by",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Submitted By"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Job_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Malformed request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "List jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Job detail with progress and result",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "description": "Cancel a Job.\n\nA `queued` Job is cancelled outright; a `running` one is *marked* and stops at its next\ncheckpoint, so the response may still report `running`. That is the honest answer —\nreporting `cancelled` while the work is still burning CPU shows a freed worker slot\nthat is not free.",
        "operationId": "cancel_job_api_v1_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Request cooperative cancellation (FR-PLAT-9)",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/events": {
      "get": {
        "description": "Server-sent events, so the UI shows progress where the action was taken (`07` §6).\n\nPolled rather than pushed. A `LISTEN`/`NOTIFY` subscription would hold a database\nconnection open per viewer, and a jobs page with twenty rows open in three browsers\nexhausts the pool. Polling one row on an interval is dull and survives a restart of\nanything.\n\nThe stream closes when the Job reaches a terminal state, so a client that forgets to\nunsubscribe does not hold the connection for ever.",
        "operationId": "stream_job_events_api_v1_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "SSE stream of progress updates",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{job_id}/logs": {
      "get": {
        "description": "Oldest first — a log read out of order is not a log.\n\nOrdered by the database-assigned `seq`, not by the UUIDv7 id: ids generated within one\nmillisecond have no defined order, and every row written in a single transaction shares\nits `at` timestamp. Both produce scrambled output.",
        "operationId": "get_job_logs_api_v1_jobs__job_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_JobLogLine_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Malformed request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Authentication failed"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not permitted"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Captured log lines (FR-PLAT-10)",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}